Add batch_update to HuaweiSolarBridge.
This function replaces the  method: instead of querying a fixed set of registers,
you can now pass a list of register names that you want to receive.

Context: the DataUpdateCoordinator objects from the Home Assistant integration now keep track
of which entities are enabled (thanks to the  feature that has been added to HA).
This allows us to only request updates for the entities which are currenlty enabled in HA.

**Note**: the license of this project has been changed to AGPL 3.0

+++ src/huawei_solar/registers.py
@@ -378,7 +378,7 @@ class TimeOfUseRegisters(RegisterDefinition):
         for period in data:
             builder.add_16bit_uint(period.start_time),
             builder.add_16bit_uint(period.end_time),
-            builder.add_32bit_uint(period.electricity_price * 1000)
+            builder.add_32bit_uint(int(period.electricity_price * 1000))
 
         # pad with empty periods
         for _ in range(len(data), LG_RESU_TOU_PERIODS):
